REFACTOR: NoticeCreatable 프로토콜로 코드 중복 제거
