+++ Projects/KNData/Sources/Repository/NoticeRepositoryImpl.swift
@@ -11,7 +11,7 @@ import KNDomain
 import KNNetwork
 import KNUtility
 
-public final class NoticeRepositoryImpl: NoticeRepository {
+public final class NoticeRepositoryImpl: NoticeRepository, NoticeCreatable {
     private let dataSource: RemoteDataSource
     private let baseURL: String? = KNDataResources.bundle.noticeURL
     
@@ -118,29 +118,4 @@ public final class NoticeRepositoryImpl: NoticeRepository {
         return notice
     }
     
-}
-
-extension NoticeRepositoryImpl {
-    func createNotice(_ data: NoticeData) -> Notice {
-        let extractedCategory: any CategoryProtocol
-        
-        if let category = NoticeCategory(rawValue: data.topic) {
-            extractedCategory = category
-        } else if let category = MajorCategory(rawValue: data.topic) {
-            extractedCategory = category
-        } else {
-            extractedCategory = NoticeCategory.generalNotice
-        }
-        
-        return Notice(
-            id: data.nttID,
-            title: data.title,
-            contentUrl: data.contentURL,
-            isSummarizable: data.isContentSummary,
-            department: data.department,
-            uploadDate: data.registrationDate,
-            imageUrl: data.contentImageURL,
-            category: extractedCategory
-        )
-    }
 }
